tutorial/requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
enum34==1.1.6
falcon==1.3.0
funcsigs==1.0.2
gunicorn==19.7.1
httpie==0.9.9
idna==2.6
ipython==5.5.0
ipython-genutils==0.2.0
mock==2.0.0
msgpack-python==0.4.8
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.4.34
Pygments==2.2.0
pytest==3.2.3
python-mimeparse==1.6.0
requests==2.18.4
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
